Separated receiver and sender
